+++ config/initializers/defaults.rb
@@ -1,3 +1,2 @@
 # http://apidock.com/rails/ActiveRecord/Serialization/to_json
 ActiveRecord::Base.include_root_in_json = false
-