+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## HEAD
-* `/platforms/{{platform_id}}/sites/{{site_id}}/status` is no longer a collection. 
+* `/sites/{{site_id}}/status` is no longer a collection. 
   Instead, there is now a `nodes` key which returns a hash with `{hostname => {:soft => soft_state, :hard => hard_state, :reservations => [...]}}`
 
 * Upon creating a job or deployment, the details of the newly created job or deployment are not returned in the response. 
@@ -7,7 +7,9 @@
 
 * Deployment notifications now contain the full detail of the deployment. What was previously sent is available in the `result`.
 
-* IN job description:
+* In `/sites/{{site_id}}/jobs`, a lot more details are returned for each job.
+
+* In job description:
   `user_uid` is replaced by `user`. `user_uid` is still available but is going to be deprecated and removed in a future revision.
   `site_uid` is removed